Add is_admin to jwt-claims, restrict item-delete endpoint with admin claim

+++ application/_resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from .._models.item import ItemModel
 
 from .._static.configuration import dbLocation
@@ -48,12 +48,20 @@ class Item(Resource):
         return item.json()
 
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilage required."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
+            return {"message": "Item deleted."}, 201
+        
+        return {"message": "Item not found."}, 404
 
-        return {"message": "Item deleted"}, 204
+
